:truck: Move routes models to own bespoke submodule
