+++ acapy_agent/anoncreds/routes/revocation/credentials/models.py
@@ -1,0 +1,213 @@
+"""AnonCreds credential revocation models."""
+
+from marshmallow import fields, validate, validates_schema
+from marshmallow.exceptions import ValidationError
+
+from .....messaging.models.openapi import OpenAPISchema
+from .....messaging.valid import (
+    ANONCREDS_CRED_REV_ID_EXAMPLE,
+    ANONCREDS_CRED_REV_ID_VALIDATE,
+    ANONCREDS_REV_REG_ID_EXAMPLE,
+    ANONCREDS_REV_REG_ID_VALIDATE,
+    UUID4_EXAMPLE,
+    UUID4_VALIDATE,
+    UUIDFour,
+)
+from ....models.issuer_cred_rev_record import (
+    IssuerCredRevRecordSchemaAnonCreds,
+)
+from ...common import (
+    create_transaction_for_endorser_description,
+    endorser_connection_id_description,
+)
+
+
+class CredRevRecordQueryStringSchema(OpenAPISchema):
+    """Parameters and validators for credential revocation record request."""
+
+    @validates_schema
+    def validate_fields(self, data, **kwargs):
+        """Validate schema fields - must have (rr-id and cr-id) xor cx-id."""
+
+        rev_reg_id = data.get("rev_reg_id")
+        cred_rev_id = data.get("cred_rev_id")
+        cred_ex_id = data.get("cred_ex_id")
+
+        if not (
+            (rev_reg_id and cred_rev_id and not cred_ex_id)
+            or (cred_ex_id and not rev_reg_id and not cred_rev_id)
+        ):
+            raise ValidationError(
+                "Request must have either rev_reg_id and cred_rev_id or cred_ex_id"
+            )
+
+    rev_reg_id = fields.Str(
+        required=False,
+        validate=ANONCREDS_REV_REG_ID_VALIDATE,
+        metadata={
+            "description": "Revocation registry identifier",
+            "example": ANONCREDS_REV_REG_ID_EXAMPLE,
+        },
+    )
+    cred_rev_id = fields.Str(
+        required=False,
+        validate=ANONCREDS_CRED_REV_ID_VALIDATE,
+        metadata={
+            "description": "Credential revocation identifier",
+            "example": ANONCREDS_CRED_REV_ID_EXAMPLE,
+        },
+    )
+    cred_ex_id = fields.Str(
+        required=False,
+        validate=UUID4_VALIDATE,
+        metadata={
+            "description": "Credential exchange identifier",
+            "example": UUID4_EXAMPLE,
+        },
+    )
+
+
+class CredRevRecordResultSchemaAnonCreds(OpenAPISchema):
+    """Result schema for credential revocation record request."""
+
+    result = fields.Nested(IssuerCredRevRecordSchemaAnonCreds())
+
+
+class PublishRevocationsOptions(OpenAPISchema):
+    """Options for publishing revocations to ledger."""
+
+    endorser_connection_id = fields.Str(
+        metadata={
+            "description": endorser_connection_id_description,
+            "required": False,
+            "example": UUIDFour.EXAMPLE,
+        }
+    )
+
+    create_transaction_for_endorser = fields.Bool(
+        metadata={
+            "description": create_transaction_for_endorser_description,
+            "required": False,
+            "example": False,
+        }
+    )
+
+
+class PublishRevocationsSchemaAnonCreds(OpenAPISchema):
+    """Request and result schema for revocation publication API call."""
+
+    rrid2crid = fields.Dict(
+        required=False,
+        keys=fields.Str(metadata={"example": ANONCREDS_REV_REG_ID_EXAMPLE}),
+        values=fields.List(
+            fields.Str(
+                validate=ANONCREDS_CRED_REV_ID_VALIDATE,
+                metadata={
+                    "description": "Credential revocation identifier",
+                    "example": ANONCREDS_CRED_REV_ID_EXAMPLE,
+                },
+            )
+        ),
+        metadata={"description": "Credential revocation ids by revocation registry id"},
+    )
+    options = fields.Nested(PublishRevocationsOptions())
+
+
+class PublishRevocationsResultSchemaAnonCreds(OpenAPISchema):
+    """Result schema for credential definition send request."""
+
+    rrid2crid = fields.Dict(
+        required=False,
+        keys=fields.Str(metadata={"example": ANONCREDS_REV_REG_ID_EXAMPLE}),
+        values=fields.List(
+            fields.Str(
+                validate=ANONCREDS_CRED_REV_ID_VALIDATE,
+                metadata={
+                    "description": "Credential revocation identifier",
+                    "example": ANONCREDS_CRED_REV_ID_EXAMPLE,
+                },
+            )
+        ),
+        metadata={"description": "Credential revocation ids by revocation registry id"},
+    )
+
+
+class RevokeRequestSchemaAnonCreds(CredRevRecordQueryStringSchema):
+    """Parameters and validators for revocation request."""
+
+    @validates_schema
+    def validate_fields(self, data: dict, **kwargs) -> None:
+        """Validate fields - connection_id and thread_id must be present if notify."""
+        super().validate_fields(data, **kwargs)
+
+        notify = data.get("notify")
+        connection_id = data.get("connection_id")
+        notify_version = data.get("notify_version", "v1_0")
+
+        if notify and not connection_id:
+            raise ValidationError("Request must specify connection_id if notify is true")
+        if notify and not notify_version:
+            raise ValidationError("Request must specify notify_version if notify is true")
+
+    publish = fields.Boolean(
+        required=False,
+        metadata={
+            "description": (
+                "(True) publish revocation to ledger immediately, or (default, False)"
+                " mark it pending"
+            )
+        },
+    )
+    notify = fields.Boolean(
+        required=False,
+        metadata={"description": "Send a notification to the credential recipient"},
+    )
+    notify_version = fields.String(
+        validate=validate.OneOf(["v1_0", "v2_0"]),
+        required=False,
+        metadata={
+            "description": (
+                "Specify which version of the revocation notification should be sent"
+            )
+        },
+    )
+    connection_id = fields.Str(
+        required=False,
+        validate=UUID4_VALIDATE,
+        metadata={
+            "description": (
+                "Connection ID to which the revocation notification will be sent;"
+                " required if notify is true"
+            ),
+            "example": UUID4_EXAMPLE,
+        },
+    )
+    thread_id = fields.Str(
+        required=False,
+        metadata={
+            "description": (
+                "Thread ID of the credential exchange message thread resulting in the"
+                " credential now being revoked; required if notify is true"
+            )
+        },
+    )
+    comment = fields.Str(
+        required=False,
+        metadata={
+            "description": "Optional comment to include in revocation notification"
+        },
+    )
+    options = PublishRevocationsOptions()
+
+
+class AnonCredsRevRegIdMatchInfoSchema(OpenAPISchema):
+    """Path parameters and validators for request taking rev reg id."""
+
+    rev_reg_id = fields.Str(
+        required=True,
+        validate=ANONCREDS_REV_REG_ID_VALIDATE,
+        metadata={
+            "description": "Revocation Registry identifier",
+            "example": ANONCREDS_REV_REG_ID_EXAMPLE,
+        },
+    )
