+++ acapy_agent/anoncreds/routes/revocation/lists/routes.py
@@ -5,56 +5,18 @@ from asyncio import shield
 
 from aiohttp import web
 from aiohttp_apispec import docs, request_schema, response_schema
-from marshmallow import fields
 
 from .....admin.decorators.auth import tenant_authentication
 from .....admin.request_context import AdminRequestContext
-from .....messaging.models.openapi import OpenAPISchema
-from .....messaging.valid import ANONCREDS_REV_REG_ID_EXAMPLE, UUIDFour
 from .....storage.error import StorageNotFoundError
 from .....utils.profiles import is_not_anoncreds_profile_raise_web_exception
 from ....models.revocation import RevListResultSchema
 from ....revocation.revocation import AnonCredsRevocation, AnonCredsRevocationError
 from ....util import handle_value_error
-from ...common import (
-    create_transaction_for_endorser_description,
-    endorser_connection_id_description,
-)
 from .. import REVOCATION_TAG_TITLE
+from .models import RevListCreateRequestSchema
 
 LOGGER = logging.getLogger(__name__)
-
-
-class RevListOptionsSchema(OpenAPISchema):
-    """Parameters and validators for revocation list options."""
-
-    endorser_connection_id = fields.Str(
-        metadata={
-            "description": endorser_connection_id_description,
-            "example": UUIDFour.EXAMPLE,
-        },
-        required=False,
-    )
-    create_transaction_for_endorser = fields.Bool(
-        metadata={
-            "description": create_transaction_for_endorser_description,
-            "example": False,
-        },
-        required=False,
-    )
-
-
-class RevListCreateRequestSchema(OpenAPISchema):
-    """Request schema for revocation registry creation request."""
-
-    rev_reg_def_id = fields.Str(
-        metadata={
-            "description": "Revocation registry definition identifier",
-            "example": ANONCREDS_REV_REG_ID_EXAMPLE,
-        },
-        required=True,
-    )
-    options = fields.Nested(RevListOptionsSchema)
 
 
 @docs(
