+++ acapy_agent/anoncreds/routes/revocation/tails/routes.py
@@ -2,34 +2,16 @@
 
 from aiohttp import web
 from aiohttp_apispec import docs, match_info_schema, response_schema
-from marshmallow import fields
 
 from .....admin.decorators.auth import tenant_authentication
 from .....admin.request_context import AdminRequestContext
 from .....anoncreds.issuer import AnonCredsIssuerError
 from .....anoncreds.routes.revocation import AnonCredsRevocationModuleResponseSchema
-from .....messaging.models.openapi import OpenAPISchema
-from .....messaging.valid import (
-    ANONCREDS_REV_REG_ID_EXAMPLE,
-    ANONCREDS_REV_REG_ID_VALIDATE,
-)
 from .....utils.profiles import is_not_anoncreds_profile_raise_web_exception
 from ....revocation.revocation import AnonCredsRevocation
 from ....util import handle_value_error
 from .. import REVOCATION_TAG_TITLE
-
-
-class AnonCredsRevRegIdMatchInfoSchema(OpenAPISchema):
-    """Path parameters and validators for request taking rev reg id."""
-
-    rev_reg_id = fields.Str(
-        required=True,
-        validate=ANONCREDS_REV_REG_ID_VALIDATE,
-        metadata={
-            "description": "Revocation Registry identifier",
-            "example": ANONCREDS_REV_REG_ID_EXAMPLE,
-        },
-    )
+from .models import AnonCredsRevRegIdMatchInfoSchema
 
 
 @docs(
